ref: moved article category schemas into schemas/article/category
