+++ schemas/article/category/ArticleCategory.ts
