+++ schemas/article/category/RawArticleCategory.ts
